MODIFY: user post 분할
