+++ kwanghwi/myapp/controllers/userController.js
@@ -34,28 +34,7 @@ const signIn = async (req, res) => {
   }
 };
 
-const posts = async (req, res) => {
-  console.log(req.userId);
-  try {
-    const { title, content, contentImage } = req.body;
-
-    if (!title || !content || !contentImage) {
-      return res.status(400).json({ message: "KEY_ERROR" });
-    }
-
-    await userService.posts(title, content, contentImage, req.userId);
-
-    console.log("req.userId", req.userId);
-
-    res.status(201).json({ message: "POSTCREATE SUCCESS" });
-  } catch (err) {
-    console.log(err);
-    return res.status(err.statusCode || 500).json({ message: err.message });
-  }
-};
-
 module.exports = {
   signUp,
   signIn,
-  posts,
 };
