+++ kwanghwi/myapp/routes/index.js
@@ -4,6 +4,8 @@ const express = require("express");
 const router = express.Router();
 
 const userRouter = require("./userRouter");
-router.use("/user", userRouter);
+const postRouter = require("./postRouter");
+router.use("/user", userRouter.router);
+router.use("", postRouter.router);
 
 module.exports = router;
