+++ kwanghwi/myapp/routes/userRouter.js
@@ -6,6 +6,7 @@ const router = express.Router();
 
 router.post("/signup", userController.signUp);
 router.post("/signin", userController.signIn);
-router.post("/posts", auth.validateToken, userController.posts);
 
-module.exports = router;
+module.exports = {
+  router,
+};
